feat: dynamic component loading
