+++ apps/web/__tests__/fixtures/newContent.json
@@ -1,5 +1,5 @@
 {
-  "name": "UiCarousel",
+  "name": "BannerCarousel",
   "options": {
     "controls": {
       "color": "#000"
