+++ apps/web/components/PageBlock/PageBlock.vue
@@ -10,7 +10,7 @@
         'mb-xl': blockSize === 'xl',
       },
       {
-        'max-w-screen-3xl mx-auto md:px-6 lg:px-10 mt-3': block.name !== 'UiCarousel',
+        'max-w-screen-3xl mx-auto md:px-6 lg:px-10 mt-3': block.name !== 'BannerCarousel',
       },
       {
         'outline outline-4 outline-[#538AEA]':
@@ -47,7 +47,7 @@
       @delete="deleteBlock"
       @change-position="changeBlockPosition"
     />
-    <component :is="getComponent && getComponent(block.name)" v-bind="block.options" :index="index" />
+    <component :is="getBlockComponent" v-bind="block.options" :index="index" />
     <button
       v-if="disableActions && isPreview"
       class="z-[0] md:z-[1] lg:z-[10] absolute bottom-0 left-1/2 transform -translate-x-1/2 translate-y-1/2 rounded-[18px] p-[6px] bg-[#538aea] text-white opacity-0 group-hover:opacity-100 group-focus:opacity-100"
@@ -74,7 +74,6 @@ interface Props {
   clickedBlockIndex: number | null;
   isTablet: boolean;
   blockHasData?: (block: Block) => boolean;
-  getComponent?: (name: string) => unknown;
   tabletEdit: (index: number) => void;
   addNewBlock: (index: number, position: number) => void;
   changeBlockPosition: (index: number, position: number) => void;
@@ -83,10 +82,29 @@ interface Props {
   deleteBlock: (index: number) => void;
 }
 
-defineProps<Props>();
+const props = defineProps<Props>();
 
 const { blockSize, drawerOpen, drawerView } = useSiteConfiguration();
 const { visiblePlaceholder } = useBlockManager();
+
+const modules = import.meta.glob('@/components/**/blocks/**/*.vue') as Record<
+  string,
+  () => Promise<{ default: unknown }>
+>;
+const getBlockComponent = computed(() => {
+  if (!props.block.name) return null;
+
+  const regex = new RegExp(`${props.block.name}\\.vue$`, 'i');
+  const matched = Object.keys(modules).find((path) => regex.test(path));
+
+  if (matched) {
+    return defineAsyncComponent({
+      loader: modules[matched],
+    });
+  }
+
+  return '';
+});
 
 const displayTopPlaceholder = (index: number): boolean => {
   const visiblePlaceholderState = visiblePlaceholder.value;
