+++ apps/web/components/blocks/BannerCarousel/Banner.vue
