+++ apps/web/components/blocks/BannerCarousel/BannerSkeleton.vue
