+++ apps/web/components/blocks/BannerCarousel/__tests__/Banner.text.spec.ts
@@ -1,10 +1,10 @@
 import { describe, it, expect } from 'vitest';
 import { mount } from '@vue/test-utils'
-import { UiBanner } from '#components';
+import { BlocksBannerCarouselBanner } from '#components';
 
 describe('Banner text', () => {
   it('should not render a pretitle if no pretitle is provided', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -27,7 +27,7 @@ describe('Banner text', () => {
   });
 
   it('should not render a title if no title is provided', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -50,7 +50,7 @@ describe('Banner text', () => {
   });
 
   it('should not render a subtitle if no subtitle is provided', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -73,7 +73,7 @@ describe('Banner text', () => {
   });
 
   it('should not render a description if no description is provided', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -96,7 +96,7 @@ describe('Banner text', () => {
   });
 
   it('should set the given text color', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -120,7 +120,7 @@ describe('Banner text', () => {
   });
 
   it('should set the given background color', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -145,7 +145,7 @@ describe('Banner text', () => {
   });
 
   it('should set the given background opacity', () => {
-    const wrapper = mount(UiBanner, {
+    const wrapper = mount(BlocksBannerCarouselBanner, {
       props: {
         bannerProps: {
           text: {
@@ -171,7 +171,7 @@ describe('Banner text', () => {
 
   describe('text alignment', () => {
     it('should align text to the left by default', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -194,7 +194,7 @@ describe('Banner text', () => {
     });
 
     it('should align text to the center', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -218,7 +218,7 @@ describe('Banner text', () => {
     });
 
     it('should align text to the right', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -244,7 +244,7 @@ describe('Banner text', () => {
 
   describe('justify content', () => {
     it('should justify content to the start by default', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -267,7 +267,7 @@ describe('Banner text', () => {
     });
 
     it('should justify content to the center', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -291,7 +291,7 @@ describe('Banner text', () => {
     });
 
     it('should justify content to the end', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -317,7 +317,7 @@ describe('Banner text', () => {
 
   describe('align items', () => {
     it('should align items to the start by default', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -340,7 +340,7 @@ describe('Banner text', () => {
     });
 
     it('should align items to the center', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
@@ -364,7 +364,7 @@ describe('Banner text', () => {
     });
 
     it('should align items to the end', () => {
-      const wrapper = mount(UiBanner, {
+      const wrapper = mount(BlocksBannerCarouselBanner, {
         props: {
           bannerProps: {
             text: {
