+++ apps/web/components/blocks/NewsletterSubscribe/NewsletterSubscribeForm.vue
@@ -86,7 +86,7 @@
 
 <script setup lang="ts">
 import { SfInput, SfTextarea, SfSwitch } from '@storefront-ui/vue';
-import type { NewsletterSubscribeProps } from '~/components/NewsletterSubscribe/types';
+import type { NewsletterSubscribeProps } from './types';
 
 const textGroup = ref(true);
 const buttonGroup = ref(true);
