+++ apps/web/components/blocks/TextCard/types.ts
