+++ apps/web/composables/useHomepage/homepageTemplateDataEn.json
@@ -1,7 +1,7 @@
 {
   "blocks": [
     {
-      "name": "UiCarousel",
+      "name": "BannerCarousel",
       "options": {
         "controls": {
           "color": "#a39f9f"
@@ -69,7 +69,7 @@
       }
     },
     {
-      "name": "UiTextCard",
+      "name": "TextCard",
       "options": {
         "text": {
           "htmlDescription": "<p>Welcome to our shop, where tech enthusiasts and fashion lovers alike will find exactly what they’re looking for! From premium headphones and high-performance drones to stylish clothing, we offer everything to make your life smarter and your style unique.</p><p>Our mission is to provide you with a carefully curated selection of products that combine quality, innovation, and design. Whether you’re in search of a new speaker for perfect sound or a casual hoodie to complete your look – we’ve got you covered.</p>",
@@ -85,14 +85,9 @@
         }
       }
     },
+
     {
-      "name": "ProductRecommendedProducts",
-      "options": {
-        "categoryId": "73"
-      }
-    },
-    {
-      "name": "UiImageText",
+      "name": "ImageText",
       "options": {
         "text": {
           "htmlDescription": "<p>Our headphone collection sets new standards in audio precision. With deep bass, crystal-clear highs, and an impressive soundstage, these headphones deliver a unique listening experience for every music genre. Combining modern design, maximum comfort, and cutting-edge technology, they are the perfect choice for anyone who refuses to compromise on sound quality.</p><ul class='list-disc pl-4 mt-4 space-y-1'><li>Premium, studio-quality sound</li><li>Comfortable fit for extended listening</li><li>Long-lasting battery life</li><li>Seamless wireless connectivity</li></ul>",
@@ -116,18 +111,7 @@
         }
       }
     },
-    {
-      "name": "ProductRecommendedProducts",
-      "options": {
-        "text": {
-          "pretitle": "More than just technology",
-          "title": "Explore our fashion collection",
-          "subtitle": "Sportive fashion and accessoires for everybody.",
-          "htmlDescription": "<a class='underline' href='https://plentyshop.plentymarkets.com/wear' target='_self'>To our full collection.</a>"
-        },
-        "categoryId": "49"
-      }
-    },
+
     {
       "name": "NewsletterSubscribe",
       "options": {
