+++ apps/web/composables/useHomepage/useHomepage.ts
@@ -1,7 +1,7 @@
 import homepageTemplateDataEn from './homepageTemplateDataEn.json';
 import homepageTemplateDataDe from './homepageTemplateDataDe.json';
 import type { HomepageData, UseHomepageDataReturn, UseHomepageDataState, SetIndex, ActiveSlideIndex } from './types';
-import type { BannerProps } from '~/components/ui/Banner/types';
+import { BannerProps } from '~/components/blocks/BannerCarousel/types';
 
 const useLocaleSpecificHomepageTemplate = (locale: string) =>
   locale === 'de' ? (homepageTemplateDataDe as HomepageData) : (homepageTemplateDataEn as HomepageData);
